Enemies and Ship Parts
Added objects that act as enemies and ship parts.

The player can send their Pikmin to defeat the enemies to bring back to the Pikmin spawner to grow more Pikmin based on the enemies health.

Added ship parts that need to be retrieved by the player in order to complete the level. These ship parts have a required amount Pikmin needed in order to retrieve.